adding separate edit avatar component
