+++ client/package.json
@@ -12,12 +12,16 @@
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
     "@lottiefiles/react-lottie-player": "^2.2.2",
+    "@material-ui/core": "^4.11.4",
+    "@material-ui/icons": "^4.11.2",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "avataaars": "^1.2.2",
     "axios": "^0.21.1",
     "bootstrap": "^4.5.2",
     "email-validator": "^2.0.4",
+    "hover.css": "^2.3.2",
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
     "react-chessground": "^1.5.0",
