+++ client/src/components/profile/profile.js
@@ -6,7 +6,10 @@ import Spinner from 'react-bootstrap/Spinner';
 import AuthService from '../../services/auth.service';
 import configAPI from '../../configs/api.config';
 import EditProfilePopup from './editprofile'
+import EditAvatarPopup from './editavatar';
 import './profile.css';
+import Avatar from 'avataaars';
+import EditIcon from '@material-ui/icons/Edit';
 
 
 const API_URL = configAPI();
@@ -21,6 +24,7 @@ class Profile extends Component {
       isProfileLoading: true,
       profileNotFound: false,
       isEditProfilePopupActive: false,
+      isEditAvatarPopupActive: false,
       profileData: {},
     };
   }
@@ -35,10 +39,19 @@ class Profile extends Component {
     });
   };
 
+  onClickEditAvatar = (e) => {
+    this.setState({
+      isEditAvatarPopupActive: true,
+    });
+   
+
+  }
+
   onClickDiscard = (e) => {
     e.preventDefault();
     this.setState({
       isEditProfilePopupActive: false,
+      isEditAvatarPopupActive: false,
     });
   };
 
@@ -65,8 +78,8 @@ class Profile extends Component {
     const username = this.props.match.params.profileId;
     try {
       const { data } = await this.fetchProfile(username);
-      console.log(data)
       const { avatar, country, email, fullname, gender, joined } = data;
+     
       this.setState({
         profileData: {
           username,
@@ -139,11 +152,35 @@ class Profile extends Component {
               </svg>
               <div className="profile-page__top-section">
                 <div>
-                  <img
-                    className="profile-page__user-img"
-                    src={`https://avataaars.io/?avatarStyle=Circle&topType=${avatar.top}&accessoriesType=${avatar.accessories}&hairColor=${avatar.hairColor}&facialHairType=${avatar.facialColor}&clotheType=${avatar.clothes}&eyeType=${avatar.eyes}&eyebrowType=${avatar.eyebrow}&mouthType=${avatar.mouth}&skinColor=${avatar.skin}&clotheColor=${avatar.clothColor}`}
-                    alt="your current profile"
+                  
+                <Avatar
+                    style={{ 
+                      display: 'inline-block',
+                      width: '160px',
+                      height: '160px',
+                      boxShadow: '0 2 4 0 #16191f',
+                      borderRadius: '50%'
+                    }}
+                    avatarStyle="Circle"
+                    topType={avatar.top}
+                    accessoriesType={avatar.accessories}
+                    hairColor={avatar.hairColor}
+                    facialHairType={avatar.facialColor}
+                    clotheType={avatar.clothes}
+                    eyeType={avatar.eyes}
+                    eyebrowType={avatar.eyebrow}
+                    mouthType={avatar.mouth}
+                    skinColor={avatar.skin}
                   />
+                  {currentUser &&
+                    currentUser.username === this.props.match.params.profileId && (
+                    <EditIcon 
+                    fontSize='large'
+                    className='profile-page__editAvatarIcon hvr-buzz-out'   
+                    onClick={this.onClickEditAvatar}
+                    />
+                  )}
+                  
                 </div>
                 <div className="profile-page__immutable-fields">
                   <div className="profile-page__field">
@@ -191,12 +228,14 @@ class Profile extends Component {
             </div>
             {
               <CSSTransition
-                in={this.state.isEditProfilePopupActive}
+                in={this.state.isEditProfilePopupActive || this.state.isEditAvatarPopupActive}
                 classNames="edit-popup"
                 timeout={200}
                 unmountOnExit
               >
-                <EditProfilePopup
+                <div>
+                  {this.state.isEditProfilePopupActive &&
+                  <EditProfilePopup
                   onClickSave={this.onClickSave}
                   onClickDiscard={this.onClickDiscard}
                   currentUser={currentUser}
@@ -205,6 +244,21 @@ class Profile extends Component {
                   }
                   {...dataForEditProfile}
                 />
+                }
+
+                  {this.state.isEditAvatarPopupActive &&
+                    <EditAvatarPopup
+                    onClickSave={this.onClickSave}
+                    onClickDiscard={this.onClickDiscard}
+                    currentUser={currentUser}
+                    rerenderProfileWithUpdatedData={
+                      this.rerenderProfileWithUpdatedData
+                    }
+                    {...dataForEditProfile}
+                  />
+                  }
+                  
+                </div>
               </CSSTransition>
             }
           </div>
